vibe_prompt_reco_vector_v0.2/requirements.txt
streamlit==1.32.0
langchain==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.24.0
torch>=2.0.0 
scikit-learn==1.3.2
faiss-cpu==1.7.4
numpy==1.24.0
torch==2.0.0
streamlit==1.32.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
faiss-cpu==1.7.4
numpy==1.24.0
torch==2.0.0    
huggingface_hub==0.25.0
supabase==2.3.5